Refactor org.biojava.nbio.core.sequence.template.SequenceMixin
